+++ iftttie/core.py
@@ -5,25 +5,25 @@ from concurrent.futures import CancelledError
 from contextlib import suppress
 
 from aiohttp import ClientConnectorError
+from aiohttp.web import Application
 from loguru import logger
 
-from iftttie import web
 from iftttie.channels.base import BaseChannel
 
 
-async def run_channels(app: web.Application):
+async def run_channels(app: Application):
     """Run channels from the configuration."""
 
     logger.info('Running channels…')
     with suppress(CancelledError):
-        await gather(*[run_channel(app, channel) for channel in app.context.channels])
+        await gather(*[run_channel(app, channel) for channel in app['context'].channels])
 
 
-async def run_channel(app: web.Application, channel: BaseChannel):
+async def run_channel(app: Application, channel: BaseChannel):
     while True:
         logger.info('Running {channel}…', channel=channel)
         try:
-            await channel.run(app.context)
+            await channel.run(app['context'])
         except CancelledError:
             logger.info('Stopped channel {channel}.', channel=channel)
             break