+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **Change: #54 `Event.key` is changed to `Event.channel_id`** 
 - New: #17 automatically refresh page every minute
 - Fix: #64 Nest image URLs should not be logged
+- Chore: do not subclass `aiohttp.web.Application`
 
 ## `0.10.0`